requirements.txt
django == 1.11
django
matplotlib
pyserial==3.4
bootstrap-admin
